Move fitLoop and fitTensors into the LayersModel class
This change removes passing the model as type `any` into both of these
functions (to avoid a circular dependency).

Alternatively, if we want these functions to remain separate, we can write a
TrainableModel interface in a separate file and use that instead of `any`.
