+++ tfjs-layers/src/engine/training.ts
@@ -14,7 +14,7 @@ import * as tfc from '@tensorflow/tfjs-core';
 import {io, ModelPredictConfig as ModelPredictArgs, NamedTensorMap, Optimizer, Scalar, scalar, serialization, Tensor, Tensor1D, tensor1d, util} from '@tensorflow/tfjs-core';
 
 import * as K from '../backend/tfjs_backend';
-import {History, ModelLoggingVerbosity} from '../base_callbacks';
+import {BaseCallback, configureCallbacks, History, ModelLoggingVerbosity, standardizeCallbacks} from '../base_callbacks';
 import {nameScope} from '../common';
 import {NotImplementedError, RuntimeError, ValueError} from '../errors';
 import {Shape} from '../keras_format/common';
@@ -22,6 +22,7 @@ import {LossIdentifier} from '../keras_format/loss_config';
 import {OptimizerSerialization} from '../keras_format/optimizer_config';
 import {MetricsIdentifier, TrainingConfig} from '../keras_format/training_config';
 import {deserialize} from '../layers/serialization';
+import { disposeTensorsInLogs, UnresolvedLogs } from '../logs';
 import * as losses from '../losses';
 import * as Metrics from '../metrics';
 import * as optimizers from '../optimizers';
@@ -39,7 +40,7 @@ import {Dataset} from './dataset_stub';
 import {execute, FeedDict} from './executor';
 import {DisposeResult, SymbolicTensor} from './topology';
 import {evaluateDataset, fitDataset, ModelEvaluateDatasetArgs, ModelFitDatasetArgs} from './training_dataset';
-import {checkBatchSize, disposeNewTensors, ensureTensorsRank2OrHigher, fitTensors, makeBatches, ModelFitArgs, sliceArrays, sliceArraysByIndices} from './training_tensors';
+import {checkBatchSize, disposeNewTensors, ensureTensorsRank2OrHigher, makeBatches, ModelFitArgs, sliceArrays, sliceArraysByIndices} from './training_tensors';
 import {ClassWeight, ClassWeightMap, computeWeightedLoss, standardizeClassWeights, standardizeWeights} from './training_utils';
 
 /**
@@ -1464,7 +1465,304 @@ export class LayersModel extends Container implements tfc.InferenceModel {
       x: Tensor|Tensor[]|{[inputName: string]: Tensor},
       y: Tensor|Tensor[]|{[inputName: string]: Tensor},
       args: ModelFitArgs = {}): Promise<History> {
-    return fitTensors(this, x, y, args);
+    if (this.isTraining) {
+      throw new Error(
+          'Cannot start training because another fit() call is ongoing.');
+    }
+    this.isTraining = true;
+    let inputs: Tensor[];
+    let targets: Tensor[];
+    let originalInputs: Tensor[];
+    let originalTargets: Tensor[];
+    let inputValX: Tensor|Tensor[];
+    let inputValY: Tensor|Tensor[];
+    let valX: Tensor|Tensor[];
+    let valY: Tensor|Tensor[];
+    let sampleWeights: Tensor[];
+    try {
+      const batchSize = args.batchSize == null ? 32 : args.batchSize;
+      checkBatchSize(batchSize);
+
+      // Validate user data.
+      // TODO(cais): Support sampleWeight.
+      const checkBatchAxis = false;
+      const standardizedOuts =
+          await this.standardizeUserData(
+              x, y, args.sampleWeight, args.classWeight, checkBatchAxis,
+              batchSize) as [Tensor[], Tensor[], Tensor[]];
+      inputs = standardizedOuts[0];
+      targets = standardizedOuts[1];
+      sampleWeights = standardizedOuts[2];
+
+      // Prepare validation data.
+      let doValidation = false;
+      let valIns: Tensor[];
+      if (args.validationData != null && args.validationData.length > 0) {
+        doValidation = true;
+        if (args.validationData.length === 2) {
+          // config.validationData consists of valX and valY.
+          inputValX = args.validationData[0];
+          inputValY = args.validationData[1];
+        } else if (args.validationData.length === 3) {
+          throw new NotImplementedError(
+              'validationData including sample weights is not supported yet.');
+        } else {
+          throw new ValueError(
+              `When passing validation data, it must contain 2 (valX, valY) ` +
+              `or 3 (valX, valY, valSampleWeight) items; ` +
+              `${args.validationData} is invalid.`);
+        }
+
+        const checkBatchAxis = true;
+        const valStandardized =
+            await this.standardizeUserData(
+                inputValX, inputValY, null, /** Unused sample weights. */
+                null,                       /** Unused class weights. */
+                checkBatchAxis, batchSize) as [Tensor[], Tensor[], Tensor[]];
+        valX = valStandardized[0];
+        valY = valStandardized[1];
+        valIns = valX.concat(valY);
+        // TODO(cais): Add useLearningPhase data properly.
+      } else if (
+          args.validationSplit != null && args.validationSplit > 0 &&
+          args.validationSplit < 1) {
+        doValidation = true;
+        // Porting Note: In tfjs-layers, inputs[0] is always a Tensor.
+        const splitAt =
+            Math.floor(inputs[0].shape[0] * (1 - args.validationSplit));
+        const originalBatchSize = inputs[0].shape[0];
+        valX = sliceArrays(inputs, splitAt, originalBatchSize) as Tensor[];
+        originalInputs = inputs;
+        inputs = sliceArrays(inputs, 0, splitAt) as Tensor[];
+        valY = sliceArrays(targets, splitAt, originalBatchSize) as Tensor[];
+        originalTargets = targets;
+        targets = sliceArrays(targets, 0, splitAt) as Tensor[];
+        // TODO(cais): Once sampleWeights becomes available, slice it to get
+        //   valSampleWeights.
+        valIns = valX.concat(valY);
+
+        // TODO(cais): Add useLearningPhase data properly.
+      } else if (args.validationSteps != null) {
+        doValidation = true;
+        // TODO(cais): Add useLearningPhase.
+      }
+
+      const ins = inputs.concat(targets).concat(sampleWeights);
+
+      this.checkTrainableWeightsConsistency();
+
+      // TODO(cais): Handle use_learning_phase and learning_phase?
+
+      // Porting Note: Here we see a key deviation of tfjs-layers from
+      // Keras.
+      //  Due to the imperative nature of tfjs-layers' backend (tfjs-core),
+      //  we do not construct symbolic computation graphs to embody the
+      //  training process. Instead, we define a function that performs the
+      //  training action. In PyKeras, the data (inputs and targets) are fed
+      //  through graph placeholders. In tfjs-layers, the data are fed as
+      //  function arguments. Since the function are defined below in the
+      //  scope, we don't have equivalents of PyKeras's
+      //  `_make_train_funciton`.
+      const trainFunction = this.makeTrainFunction();
+      const outLabels = this.getDedupedMetricsNames() as string[];
+
+      let valFunction: (data: Tensor[]) => Scalar[];
+      let callbackMetrics: string[];
+      if (doValidation) {
+        this.makeTestFunction();
+        valFunction = this.testFunction;
+        callbackMetrics =
+            outLabels.slice().concat(outLabels.map(n => 'val_' + n));
+      } else {
+        valFunction = null;
+        valIns = [];
+        callbackMetrics = outLabels.slice();
+      }
+
+      const callbacks = standardizeCallbacks(args.callbacks, args.yieldEvery);
+      const out = await this.fitLoop(
+          trainFunction, ins, outLabels, batchSize, args.epochs,
+          args.verbose, callbacks, valFunction, valIns, args.shuffle,
+          callbackMetrics, args.initialEpoch, null, null);
+      return out;
+    } finally {
+      this.isTraining = false;
+      // Memory clean up.
+      disposeNewTensors(inputs, x);
+      disposeNewTensors(targets, y);
+      disposeNewTensors(originalInputs, x);
+      disposeNewTensors(originalTargets, y);
+      disposeNewTensors(valX as Tensor[], inputValX);
+      disposeNewTensors(valY as Tensor[], inputValY);
+      if (sampleWeights != null) {
+        tfc.dispose(sampleWeights);
+      }
+    }
+    // TODO(cais): Add value to outLabels.
+  }
+
+  /**
+   * Abstract fit function for `f(ins)`.
+   * @param f A Function returning a list of tensors. For training, this
+   *   function is expected to perform the updates to the variables.
+   * @param ins List of tensors to be fed to `f`.
+   * @param outLabels List of strings, display names of the outputs of `f`.
+   * @param batchSize Integer batch size or `== null` if unknown. Default : 32.
+   * @param epochs Number of times to iterate over the data. Default : 1.
+   * @param verbose Verbosity mode: 0, 1, or 2. Default: 1.
+   * @param callbacks List of callbacks to be called during training.
+   * @param valF Function to call for validation.
+   * @param valIns List of tensors to be fed to `valF`.
+   * @param shuffle Whether to shuffle the data at the beginning of every
+   * epoch. Default : true.
+   * @param callbackMetrics List of strings, the display names of the metrics
+   *   passed to the callbacks. They should be the concatenation of the
+   *   display names of the outputs of `f` and the list of display names
+   *   of the outputs of `valF`.
+   * @param initialEpoch Epoch at which to start training (useful for
+   *   resuming a previous training run). Default : 0.
+   * @param stepsPerEpoch Total number of steps (batches on samples) before
+   *   declaring one epoch finished and starting the next epoch. Ignored with
+   *   the default value of `undefined` or `null`.
+   * @param validationSteps Number of steps to run validation for (only if
+   *   doing validation from data tensors). Not applicable for tfjs-layers.
+   * @returns A `History` object.
+   */
+  async fitLoop(
+      f: (data: Tensor[]) => Scalar[], ins: Tensor[], outLabels?:
+      string[], batchSize?: number, epochs?: number, verbose?: number,
+      callbacks?: BaseCallback[], valF?: (data: Tensor[]) => Scalar[], valIns?:
+      Tensor[], shuffle?: boolean|string, callbackMetrics?: string[],
+      initialEpoch?: number, stepsPerEpoch?: number, validationSteps?: number):
+      Promise<History> {
+    if (batchSize == null) {
+      batchSize = 32;
+    }
+    if (epochs == null) {
+      epochs = 1;
+    }
+    if (shuffle == null) {
+      shuffle = true;
+    }
+    if (initialEpoch == null) {
+      initialEpoch = 0;
+    }
+
+    // TODO(cais): Change const to let below when implementing validation.
+    let doValidation = false;
+    if (valF != null && valIns != null) {
+      doValidation = true;
+      // TODO(cais): verbose message.
+    }
+    if (validationSteps != null) {
+      doValidation = true;
+      if (stepsPerEpoch == null) {
+        throw new ValueError(
+            'Can only use `validationSteps` when doing step-wise training, ' +
+            'i.e., `stepsPerEpoch` must be set.');
+      }
+    }
+
+    const numTrainSamples =
+        this.checkNumSamples(ins, batchSize, stepsPerEpoch, 'steps_per_epoch');
+    let indexArray: number[];
+    if (numTrainSamples != null) {
+      indexArray = range(0, numTrainSamples);
+    }
+
+    if (verbose == null) {
+      verbose = 1;
+    }
+
+    const {callbackList, history} = configureCallbacks(
+        callbacks, verbose, epochs, initialEpoch, numTrainSamples,
+        stepsPerEpoch, batchSize, doValidation, callbackMetrics);
+    callbackList.setModel(this);
+    this.history = history;
+    await callbackList.onTrainBegin();
+    this.stopTraining_ = false;
+    // TODO(cais): Take care of callbacks.validation_data as in PyKeras.
+    // TODO(cais): Pre-convert feeds for performance as in PyKeras.
+
+    for (let epoch = initialEpoch; epoch < epochs; ++epoch) {
+      await callbackList.onEpochBegin(epoch);
+      const epochLogs: UnresolvedLogs = {};
+      if (stepsPerEpoch != null) {
+        throw new NotImplementedError(
+            'stepsPerEpoch mode is not implemented yet.');
+      } else {
+        if (shuffle === 'batch') {
+          throw new NotImplementedError('batch shuffling is not implemneted'
+                                        + ' yet');
+        } else if (shuffle) {
+          util.shuffle(indexArray);
+        }
+        // Convert the potentially shuffled indices to Tensor1D, to avoid the
+        // cost of repeated creation of Array1Ds later on.
+        const epochIndexArray1D = tensor1d(indexArray);
+
+        const batches = makeBatches(numTrainSamples, batchSize);
+        for (let batchIndex = 0; batchIndex < batches.length; ++batchIndex) {
+          const batchLogs: UnresolvedLogs = {};
+          await callbackList.onBatchBegin(batchIndex, batchLogs);
+
+          tfc.tidy(() => {
+            const batchStart = batches[batchIndex][0];
+            const batchEnd = batches[batchIndex][1];
+            const batchIds = K.sliceAlongFirstAxis(
+                                 epochIndexArray1D, batchStart,
+                                 batchEnd - batchStart) as Tensor1D;
+            batchLogs['batch'] = batchIndex;
+            batchLogs['size'] = batchEnd - batchStart;
+
+            // TODO(cais): In ins, train flag can be a number, instead of an
+            //   Tensor? Do we need to handle this in tfjs-layers?
+            const insBatch = sliceArraysByIndices(ins, batchIds) as Tensor[];
+            const outs = f(insBatch);
+            for (let i = 0; i < outLabels.length; ++i) {
+              const label = outLabels[i];
+              const out = outs[i];
+              batchLogs[label] = out;
+              tfc.keep(out);
+              // TODO(cais): Use scope() to avoid ownership.
+            }
+
+            if (batchIndex === batches.length - 1) {  // Last batch.
+              if (doValidation) {
+                const valOuts = this.testLoop(valF, valIns, batchSize);
+                // Porting Notes: In tfjs-layers, valOuts is always an Array.
+                for (let i = 0; i < outLabels.length; ++i) {
+                  const label = outLabels[i];
+                  const out = valOuts[i];
+                  tfc.keep(out);
+                  // TODO(cais): Use scope() to avoid ownership.
+                  epochLogs['val_' + label] = out;
+                }
+              }
+            }
+          });
+
+          await callbackList.onBatchEnd(batchIndex, batchLogs);
+          disposeTensorsInLogs(batchLogs);
+
+          if (this.stopTraining_) {
+            break;
+          }
+          // TODO(cais): return outs as list of Tensor.
+        }
+
+        epochIndexArray1D.dispose();
+      }
+      // TODO(cais): Run validation at the end of the epoch.
+      await callbackList.onEpochEnd(epoch, epochLogs);
+      if (this.stopTraining_) {
+        break;
+      }
+    }
+    await callbackList.onTrainEnd();
+
+    await this.history.syncData();
+    return this.history;
   }
 
   // TODO(cais): Add code snippet below when it's possible to instantiate
